Move GetTempPath() to testutil
Move GetTempPath() to testutil. (Mustafa)
Add a source file for unit test utils. (Mustafa)
