+++ src/Makefile.test.include
@@ -85,6 +85,8 @@ BITCOIN_TESTS =\
   test/streams_tests.cpp \
   test/test_absolute.cpp \
   test/test_absolute.h \
+  test/testutil.cpp \
+  test/testutil.h \
   test/timedata_tests.cpp \
   test/transaction_tests.cpp \
   test/txvalidationcache_tests.cpp \
